Correção de bugs no modulo de impostos e ajustes no modulo dark

+++ src/routes/web.php
@@ -216,13 +216,13 @@ Route::middleware([
     });
 
     Route::prefix('taxes')->name('taxes.')->group(function () {
-        Route::get('/', [TaxRuleController::class, 'index'])->name('index');
-        Route::get('/data', [TaxRuleController::class, 'data'])->name('data');
-        Route::get('/create', [TaxRuleController::class, 'create'])->name('create');
-        Route::post('/', [TaxRuleController::class, 'store'])->name('store');
-        Route::get('/{rule}', [TaxRuleController::class, 'edit'])->name('edit');
-        Route::put('/{rule}', [TaxRuleController::class, 'update'])->name('update');
-        Route::delete('/{rule}', [TaxRuleController::class, 'destroy'])->name('destroy');
+        Route::get('/', [TaxRuleController::class, 'index'])->name('index')->middleware('check.permission:view_post,taxas');
+        Route::get('/data', [TaxRuleController::class, 'data'])->name('data')->middleware('check.permission:view_post,taxas');
+        Route::get('/create', [TaxRuleController::class, 'create'])->name('create')->middleware('check.permission:create_post,taxas');
+        Route::post('/', [TaxRuleController::class, 'store'])->name('store')->middleware('check.permission:create_post,taxas');
+        Route::get('/{rule}', [TaxRuleController::class, 'edit'])->name('edit')->middleware('check.permission:edit_post,taxas');
+        Route::put('/{rule}', [TaxRuleController::class, 'update'])->name('update')->middleware('check.permission:edit_post,taxas');
+        Route::delete('/{rule}', [TaxRuleController::class, 'destroy'])->name('destroy')->middleware('check.permission:edit_post,taxas');
     });
 });
